requirements.txt
pyyaml==6.0.1
python-frontmatter==1.0.1
pybtex==0.24.0
setuptools==69.0.3